backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlmodel==0.0.8
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
celery==5.2.3
redis==4.1.0
alembic==1.7.5
python-dotenv==0.19.2
pytest==6.2.5
httpx==0.19.0
pyotp==2.6.0
email-validator==1.1.3
python-slugify==5.0.2
Pillow==8.4.0 